chore(PoiMap): rename PoiGeoJson to PoiMap
