+++ packages/visualizations-react/src/index.tsx
@@ -5,4 +5,4 @@ export { default as KpiCard } from './components/KpiCard';
 export { default as ChoroplethGeoJson } from './components/ChoroplethGeoJson';
 export { default as ChoroplethVectorTiles } from './components/ChoroplethVectorTiles';
 export { default as ChoroplethSvg } from './components/ChoroplethSvg';
-export { default as PoiGeoJson } from './components/PoiGeoJson';
+export { default as PoiMap } from './components/PoiMap';
