+++ packages/visualizations-react/stories/Poi/PoiMap.stories.tsx
@@ -3,8 +3,8 @@ import { BBox } from 'geojson';
 import { PoiMapData } from '@opendatasoft/visualizations';
 import { ComponentMeta, ComponentStory } from '@storybook/react';
 
-import { PoiGeoJson } from '../../src';
 import { shapes as data } from './data';
+import { PoiMap } from '../../src';
 
 const BASE_STYLE = 'https://demotiles.maplibre.org/style.json';
 
@@ -21,14 +21,14 @@ const layers : PoiMapData["layers"] = [{
 
 const bbox : BBox = [ -6.855469,41.343825,11.645508,51.371780];
 
-const meta: ComponentMeta<typeof PoiGeoJson> = {
-    title: 'Poi/GeoJson',
-    component: PoiGeoJson,
+const meta: ComponentMeta<typeof PoiMap> = {
+    title: 'Poi/PoiMap',
+    component: PoiMap,
 };
 
 export default meta;
 
-const Template: ComponentStory<typeof PoiGeoJson> = args => (
+const Template: ComponentStory<typeof PoiMap> = args => (
     <div
         style={{
             width: '50%',
@@ -38,14 +38,14 @@ const Template: ComponentStory<typeof PoiGeoJson> = args => (
             border: '1px solid black',
         }}
     >
-        <PoiGeoJson {...args} />
+        <PoiMap {...args} />
     </div>
 );
 
 /**
  * STORY: No layer params
  */
-export const PoiMapNoLayersParams : ComponentStory<typeof PoiGeoJson> = Template.bind({});
+export const PoiMapNoLayersParams : ComponentStory<typeof PoiMap> = Template.bind({});
 const PoiMapNoLayersParamsArgs = {
     data: {},
     options: {style: BASE_STYLE, bbox}
@@ -55,7 +55,7 @@ PoiMapNoLayersParams.args = PoiMapNoLayersParamsArgs;
 /**
  * STORY: No interactive
  */
-export const PoiMapNonInteractive : ComponentStory<typeof PoiGeoJson> = Template.bind({});
+export const PoiMapNonInteractive : ComponentStory<typeof PoiMap> = Template.bind({});
 const PoiMapNonInteractiveArgs = {
     data:  {value:{ layers, sources: { [layers[0].source] : {type: "geojson" as const, data}}}},
     options: {
@@ -69,7 +69,7 @@ PoiMapNonInteractive.args = PoiMapNonInteractiveArgs;
 /**
  * STORY: With match expression
  */
-export const PoiMapMatchExpression : ComponentStory<typeof PoiGeoJson> = Template.bind({});
+export const PoiMapMatchExpression : ComponentStory<typeof PoiMap> = Template.bind({});
 const PoiMapMatchExpressionArgs = {
     data: {
         value: {
