+++ packages/visualizations/src/components/MapPoi/Map.svelte
