+++ packages/visualizations/src/index.ts
@@ -3,7 +3,7 @@ export { default as MarkdownText } from './components/MarkdownText';
 export { default as KpiCard } from './components/KpiCard';
 export { ChoroplethGeoJson, ChoroplethVectorTiles } from './components/Map/WebGl';
 export { default as ChoroplethSvg } from './components/Map/Svg';
-export { default as PoiGeoJson } from './components/MapPoi';
+export { default as PoiMap } from './components/MapPoi';
 
 export * from './types';
 export * from './components/types';
